requirements.txt
python-dotenv
langchain
openai
huggingface_hub
google-generativeai
langchain_google_genai
streamlit
chainlit
pandas
openpyxl
faker
langchain_experimental
langchain-community
tiktoken
faiss-cpu
chromadb
PyPDF2
docx2txt
sentence-transformers
InstructorEmbedding
six
matplotlib
sympy
fastapi[all]
uvicorn[standard]
